+++ StoryCom/src/main/java/com/storycom/controllers/HomeController.java
@@ -6,6 +6,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class HomeController extends Base {
@@ -16,7 +17,42 @@ public class HomeController extends Base {
 
     @GetMapping(value = "/")
     public String home(Model model) {
+
         model.addAttribute("globalMenu", GLOBAL_MENU);
-        return "home";
+        model.addAttribute("submenu", "home");
+
+        return "home/home";
+    }
+
+    @GetMapping(value = "/about")
+    public String about(Model model) {
+
+        model.addAttribute("globalMenu", GLOBAL_MENU);
+        model.addAttribute("submenu", "about");
+
+        return "home/about";
+    }
+
+    @GetMapping(value = "/contact")
+    public String prepareContact(Model model) {
+
+        model.addAttribute("globalMenu", GLOBAL_MENU);
+        model.addAttribute("submenu", "contact");
+
+        return "home/contact";
+    }
+
+    @PostMapping(value = "/contact")
+    public String contactUs(Model model) {
+        return "redirect:/about/contact";
+    }
+
+    @GetMapping(value = "/timeline")
+    public String timeline(Model model) {
+
+        model.addAttribute("globalMenu", GLOBAL_MENU);
+        model.addAttribute("submenu", "timeline");
+
+        return "home/timeline";
     }
 }
