+++ StoryCom/src/main/resources/messages/messages_en.properties
@@ -1,8 +1,8 @@
 
-title.home = Home
-title.about.general = About
-title.about.contact = Contact
-title.about.timeline = Timeline
+title.home.general = Home
+title.home.about = About
+title.home.contact = Contact
+title.home.timeline = Timeline
 title.login = Log in
 title.signup = Sign up
 title.admin = Admin
@@ -21,10 +21,10 @@ title.error.pnf404 = Page not found
 title.error.mnf405 = Method not found
 title.error.ise500 = Internal server error
 
-nav.home = Home
-nav.about.general = About
-nav.about.contact = Contact
-nav.about.timeline = Timeline
+nav.home.general = Home
+nav.home.about = About
+nav.home.contact = Contact
+nav.home.timeline = Timeline
 nav.login = Login
 nav.signup = Sign up
 nav.logout = Logout
