+++ StoryCom/src/main/webapp/WEB-INF/pages/fragments/mainMenu.html
@@ -109,16 +109,17 @@
                                         <i class="fa fa-home nav-link-icon"></i>
                                     </a>
                                     <div class="dropdown-menu">
+                                        <a class="dropdown-item" th:classappend="${submenu eq 'home'} ? 'active'"
+                                           href="#" th:href="@{/}" th:text="#{nav.home.general}">Home</a>
+
                                         <a class="dropdown-item" th:classappend="${submenu eq 'about'} ? 'active'"
-                                           href="#" th:href="@{/about}" th:text="#{nav.about.general}">About</a>
+                                           href="#" th:href="@{/about}" th:text="#{nav.home.about}">About</a>
 
                                         <a class="dropdown-item" th:classappend="${submenu eq 'contact'} ? 'active'"
-                                           href="#" th:href="@{/about/contact}"
-                                           th:text="#{nav.about.contact}">Contact</a>
+                                           href="#" th:href="@{/contact}" th:text="#{nav.home.contact}">Contact</a>
 
                                         <a class="dropdown-item" th:classappend="${submenu eq 'timeline'} ? 'active'"
-                                           href="#" th:href="@{/about/timeline}"
-                                           th:text="#{nav.about.timeline}">Timeline</a>
+                                           href="#" th:href="@{/timeline}" th:text="#{nav.home.timeline}">Timeline</a>
                                     </div>
                                 </li>
                                 <li class="nav-item dropdown">
