+++ StoryCom/src/main/webapp/WEB-INF/pages/home/about.html
@@ -4,16 +4,17 @@
       xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
       layout:decorate="~{templates/contentTemplate}">
 <head>
-    <title th:text="#{title.about.general} + ' | ' + #{app.title}"></title>
+    <title th:text="#{title.home.general} + ' | ' + #{app.title}"></title>
 </head>
 <body>
 <div layout:fragment="contentDiv">
     <div class="row">
-        <div class="col-md-3 order-md-1 order-2">
-            <div th:insert="about/submenu :: #submenuDiv"></div>
-        </div>
+        <!--<div class="col-md-3 order-md-1 order-2">-->
+            <!--<div th:insert="home/submenu :: #submenuDiv"></div>-->
+        <!--</div>-->
 
-        <div class="col-md-9 order-md-2 order-1">
+        <!--<div class="col-md-9 order-md-2 order-1">-->
+        <div class="col-12">
             <p>About page</p>
         </div>
     </div>
